AUT-3779: Add utility to retrieve a channel-specific template
It has been decided that the Strategic App will reuse the
existing Web journey by replaying journeys created for the Web in
mobile WebViews. The ADR[^1] describes two technical approaches
to support this, one of which is 'Independent Pages' (i.e. where
there is significant difference between the Web and Mobile
presentation of a journey step, this will be reflected in there
being separate templates).

The utility added in this commit will support the 'Independent
Pages' approach by allowing the template to be selected based
on the channel and a mapping between Web templates and their
mobile equivalents.

[^1]: https://github.com/govuk-one-login/architecture/pull/738