AUT-3779: Introduce locale section for mobile-specific pages
The decision to replay the Web journey within mobile apps could
present a maintainability challenge for developers. My sense is that
a clear way to identify the channel for content will help future
developers, and will be much easier than mixing Web and Mobile
content.

With this in mind, this commit introduces a new top-level property
to translation files called 'mobileAppPages' that is intended to hold
all content for pages that are specific to mobile.

The new block has been populated with content for all changes under
AUT-3779.