AUT-3779: Extract template selection logic to function with tests
Introduces the getAccountNotFoundWebTemplate function which replicates
existing template selection logic previously included in the controller

I've opted to do this because the additional requirement to select a
template that is appropriate for the channel will introduce additional
logic. My sense is that two functions (one that selects the web template
and another that gets the mobile equivalent of that template) will be
easier to understand and test.