AUT-3779: Extract template selection logic to function with tests
Introduces the getAccountNotFoundTemplate function which replicates
existing template selection logic previously included in the controller
and calls getChannelSpecificTemplate.

I've opted to do this because the additional requirement to select a
template that is appropriate for the channel will introduce additional
logic. My sense is that two functions (one that selects the web template
and calls another to retrieve the mobile equivalent of that template) is
easier to understand and test.