AUT-3780: Limit max node connections
Due to the nature of nodeJS, it is capable of handling and accepting more requests than it is able to fulfil. This can be caused through a variety of reasons, and manifests itself as increased latency in responses. We can observe thisas eventLoopLag.

In a ECS context, if we allow containers to receiving too many connections, then they will just slow down and block on i/o rather than CPU - meaning that they do not report themselves as being busy against either default scaling metrics (CPUUtilized and MemUtilized).

Overload-protection is being used here to to manage to eventLoopLag. Should the eventLoop start to lag under the weight of too many connections, reject the connections with a 503 error. This prevents the container getting too busy, and aside from downstream network i/o requests made by your business logic, will stop your express apps from locking up.

The config for overload-protection in this commit was taken from here: https://github.com/govuk-one-login/di-account-management-frontend/pull/1710
This will likely be monitored and adjusted over time. In this commit, we are deploying only to staging. I have set the logging to info so that we can
see how it works when testing. This may be turned off after testing. Logging will always be turned off in production.