AUT-3762: Extract server start logic to tested method
Logic will be added soon that will gracefully shutdown the server on SIGTERM. To ensure there is no regression, this should be unit tested.

Before doing that, make the existing start server logic testable and add tests to assert the server starts on the expected port, with the expected timeouts and with the vital-signs package.