AUT-3762: Maintain one redis client with disconnect
An upcoming commit will add graceful shutdown to our server. The redis client can be disconnected, but in order to do so the client needs to be available outside of the method that creates it.

A new `disconnectRedisClient()` method is added that will disconnect the client if there is a client set.

New calls to `getSessionStore` will reuse an existing redis client so long as the redisConfig matches. Otherwise an error is thrown.