BAU: Set a default session expiry value
Set a default session expiry value to 1 hour. Express is updated in he following commit. This fixed a bug where invalid expiry dates for cookies throw an error. In our integration tests on GitHub Actions the tests will fail without this (as we dont set the env var explicitally in these tests).