BAU: Group all dependabot update npm patch dependencies
As patch bumps are the least likely to impact our use of a dependency, they should be a relativly safe initial grouping to attempt to lower the number of dependabot PRs we need to rebase/review/merge.

The current problem is we get daily PRs for patch bumps, and they end up being out of date with the `main` as a previous dependabot PR merges. This means it's currently a lengthy process of:

- Dep A: approve
- Dep B: approve
- Dep C: approve
- Dep A: merged
- Dep B: out of date
- Dep C: out of date
- Dep B: rebase
- Dep C: rebase
- Dep B: approve
- Dep B: merged
- Dep C: out of date
- Dep C: rebase
- Dep C: approve
- Dep C: merged