AUT-3838: Use channel template for SMS OTP requests exceeded
Updates the security code error controller and associated tests
to ensure a channel specific template is used when too many
requests have been made for a SMS OTP during a password reset
journey.

This is necessary because content that is appropriate to the
Web is not relevant to the Strategic App channel.