AUT-3838: Scenario 1 - Email OTP
Introduces the mobile specific "You asked to resend the security code
too many times" templates for the scenario where a user has requested
a new security code 6 times on the "Check your email" screen. Also
introduces a mobile specific variant of `index-wait.njk` because this
template is shown when a user interacts with the language toggle.