BAU: Reflect changes to global flags
The GOV.UK Frontend Analytics v3 flips the disable flags to be
enable flags to address some confusion (for example, the
isGa4Disabled flag in v2 is now a isGa4Enabled flag). Read more
about this change in Confluence[^1]

This commit updates the codebase to reflect this switch by
replacing functions and variables to be their equivalent when
flipped. For example:

- the googleAnalytics4Disabled() function which used the
  GA4_DISABLED environment variable and defaulted to "true"
  is now a googleAnalytics4Enabled() function that uses
  GA4_ENABLED and defaults to "false"
- where 'ua_disabled = false' was, 'ua_enabled = true' now is

[^1]: https://govukverify.atlassian.net/wiki/spaces/DIFC/pages/3801317710/Implementation+Guide#%E2%AC%86%EF%B8%8F-Upgrading-from-v2-to-v3