BAU: Add integration test for cookies post
The test itself is very basic, just testing that we can successfully make a request to the post endpoint (without testing any of the expected functionality, which is already tested and much easier to test in the unit tests). This, among other breaking changes, prevents breaking updates to the cookies library. This has recently been seen in express dependency bumps, which bumps the cookie library to 0.7.x, which contains extra regex validation of the cookie values, which we don't currently meet as we use double quotes in the value.

This should prevent re-occurences of this incident