BAU: Add regression test for authorize controller
The test itself is very basic, just testing that we can successfully make a request to the endpoint when the user has given cookie consent (without testing any of the expected functionality, which is already tested and much easier to test in the unit tests). This, among other breaking changes, prevents breaking updates to the cookies library, since in the case where this value is true, and the client has set cookie sharing to true, we set a cookie value and if this fails we can fail to let someone through authorize. This has recently been seen in express dependency bumps, which bumps the cookie library to 0.7.x, which contains extra regex validation of the cookie values, which we don't currently meet as we use double quotes in the value.

    This should prevent re-occurences of this incident.