AUT-3824: Add type for Reverification Result Response
Note that this type only covers the happy path (where we'll get a 200 response with populated sub and success: true), and the case where there is an error in the backend (where we'll get a non 200 response). In the unhappy path (where reverification in IPV is unsuccessful), we'll have the following additional fields: failure_code and failure_description which we'll need to add to the type and handle. In this case, we'll still have a 200 response but will need to respond to these optional fields.