BAU: Replace simple chai as promised assertions with native node assert
Chai as promised is used here to assert that a promise fails with an error. We can instead replace it with native node assert, which contains this same functionality.

We want to remove chai-as-promised as it is somewhat deprecated, and its documentation [1] suggests it's only used for test runners without async support, which mocha now has.

1: https://www.npmjs.com/package/chai-as-promised#user-content-working-with-non-promisefriendly-test-runners