BAU: replace chai-as-promised in tests use eventually
This used some functionality in chai-as-promised that's a bit harder to get in nice succinct form without it. I've introduced a helper function here to tidy it up.

We want to remove chai-as-promised as it is somewhat deprecated, and its documentation [1] suggests it's only used for test runners without async support, which mocha now has.

1: https://www.npmjs.com/package/chai-as-promised#user-content-working-with-non-promisefriendly-test-runners