BAU: Start of template validation middleware
A new middleware, intended for build envs and lower, will be introduced that will validate that variables referenced in templates have a value.

This is needed because we've had incidents of undefined variables in templates being deployed to prod where they are expected to have a value.

Nunjucks does have a `throwOnUndefined` option, but unfortunately it's scope is across all templates being rendered. This means templates outside of our control (i.e. the GOV.UK Design System) is also included in checks for undefined values, of which there are lots.

Instead we only really care about undefined values in our templates, which this will validate.

For this initial commit, only a no-operation middleware is added, which wraps the responses `render` method. Later commits will hook in at this point to do the validation.