BAU: Template validation logs properties accessed
First step to validating properties have values is to collect a list of properties accessed in a template.

We do this here by asking `nunjucks` to compile the provided template (which creates JavaScript as a string), then ask `acorn` to create an AST from the compiled JavaScript string. `acorn` is the same library used by `prettier` to process JavaScript code.

`acorn-walk` is then used to evaluate the template and pull out references both translation keys and variables accessed in the template.

These are debugged logged for now.

Later commits will evaluate if the accessed properties have values, and warn if not.