BAU: Validator excludes vars set in template
The template validator should not care about access to variables that are set within the scope of the template. This is because the validator will be used to highlight were variables are accessed that have not been set in the context of the render.

`ga4OnPageLoad` is explicitally included as it's the only property so far seen that is accessed in the template but it is defined in a base template. If more of these are seen a more robust approach can be taken to dig into base templates.