BAU: Validator warns of accesses to properties with no value
Template validator now warns when the template looks up a property that is not available in the request context, either a variable or template key.

This can be used to find rendered templates with potentially incorrect values set, with the idea to fix any issues found then make the validator throw an error instead of warn.

This wll allow devs and test runs to see quickly when a variable is accessed, but has no value.