AUT-3951: Add validation for radio buttons on cannot change security codes page
This required adding a post. We're not doing anything meaningful on successful post yet (will be done in subsequent tickets) so for the moment I'm just returning a dummy page. This won't be deployed in any production environments as the router is behind a feature flag.

Text hard-coded in English for now - following commit will add translation