POC: Add script to deploy to dev environment
This script is not intended to be used unless you understand the consequences of manually deploying to the dev environment.

This script also probably isn't useful after the migration to Secure Pipelines.

Normally we deploy to dev using a pipeline. However, in some cases it's beneficial to deploy just the frontend using this script (for example if dev is the only environment configured correctly for something you need to do).

In almost all cases, it's better to deploy to an authdev or sandpit environment (these environments are capable of handling script deploys, whereas dev usually expects a pipeline to run).

This change adds a script to deploy directly to the dev frontend, similar to what we already have for authdevs and sandpit.