AUT-4005: Fix bug where valid auth app code not recognised in MFA reset journey
The journey on the user's session was ACCOUNT_RECOVERY as a result of starting the MFA reset with IPV journey. This caused the VerifyMfaCodeHandler in the back end to get an invalid auth app secret.
As this is in the case where the identity reproving with IPV has failed, the journey type should go back to what it was before initiating MFA reset with IPV (which can be reauthentication or sign_in).
As a fix, I’ve set req.session.user.isAccountRecovery to false if IPV returns an error response.