AUT-3962: Allow going back after starting MFA reset with IPV
Once a user has started the MFA reset with IPV, they are taken to IPV. Here we enable support for going back to the previous MFA entry page (be it SMS or App).

This is done by first making the `IPV_CALLBACK` path in response to `IPV_REVERIFICATION_INIT` optional. We don't actually use the output of the state machine to drive the redirect in a0bfb6478d86d7481ad9364f34980491bee9323f. This is because the actual URL to IPV goes to an external application, not something managed by the state machine.

Instead the internal path returned after `IPV_REVERIFICATION_INIT` are the MFA code entry page for the type the user has already. This is stored in the users session and if a user navigates back from IPV to auth this last page (MFA code entry) is allowed.