AUT-3957: Simplify including ga4 script
In a future commit, `@govuk-one-login/frontend-analytics` will be updated and support for 5 taxonomy levels will be added. At the moment we only support 2. Taxonomies are defined in `src/utils/taxonomy.ts` and included as variables to nunjucks by `src/middleware/get-analytics-properties-middleware.ts`. This was after a refactor where taxonomies were previously embedded in templates (which lead to other problems).

What was left over in the refactor was on each template, the `ga4OnPageLoad()` macro call required a reference to each variable as an argument of the macro. Nunjucks macros do not support referencing a variable that is not passed into the macro as an argument.

Without any changes to this, adding the 3 new taxonomy levels will add lots of duplicate new params to `ga4OnPageLoad()` calls, which is fragile and prone to being forgotten.

Instead, the macro for `ga4OnPageLoad()` is replaced with directly including the GA4 script template, preceded by setting the same `ga4Params` values as previously used when calling the macro.

This allows the removal of the unnecessary `taxonomyLevel1: taxonomyLevel1`, `taxonomyLevel2: taxonomyLevel2`, `contentId: contentId` etc references and still allows for taxonomy values to be set/overridden per template.