AUT-4052: Add description for bypass CIDR blocks
It's difficult to keep track of what each CIDR block is in
secretsmanager. Add a new variable to allow for a description.

Currently, the unique cidrs from both of these blocks are used to build
the allowlist. Once this is merged, we can remove the old variable from
terraform and secretsmanager.