AUT-4041: Redirect users to the guidance page if they are attempting mfa reset from the app
There is an issue whereby if the user starts their journey in an app, the
user will not be able to complete an MFA reset journey in app. This is due
to linking issues in the app portions of the journey.

We therefore need to detect if a user is signing in via an app and, if so, show them
a guidance page that instructs them to sign in via a web browser to complete an MFA reset.

Currently this only applies to users using the v2 one login app, although in future it
will also apply to users coming from the hmrc app.

This approach may be revised in future to allow this journey in app,
but for our MVP new journey, we are displaying guidance to the user.

Adds an integration test that tests this redirect, by first hitting the
mfa reset page, and then hitting the route that this redirects to.

To do this, we've used the supertest agent as it maintains the session between calls.
Modifying the sinon stub to only update when the session for the first page
requested and skipping for future calls allows setup.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>