AUT-4050: Set analytics cookie preferences to off for strategic app journeys
For strategic app journeys:
* We don't want to display the cookie banner
* We want to disable analytics cookies, regardless of the user's preferences (without affecting what happens if they go through a web journey)

This achieves this by setting analytics cookie preferences to false in the authorize controller (which is the start of a user's journey). This will mean that the banner will not appear for a user within this journey (as we only display when this cookie is not set) and also that no analytics data will be sent during this journey.