AUT-4081: Check & persist allowed optional paths state for /enter-code
When a user is on the `/enter-code` page, starts the MFA reset journey and hits back they go back to the `/enter-code` page. The user is then unable to resend the code as the state machine is not in the right state to do so.

`allowAndPersistUserJourneyMiddleware` is introduced to solve this problem. It asserts the user is allowed to be on the page. It also updates the user session `journey` to reflect the path they are on (if the path is an optionalPath from a previous page). This ensures that when the users loads the page they can do any of the intended actions of that page.

This new middleware is only applied to the `GET /enter-code` route and we can consider if we want to roll this out to more, or all, routes.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>