BAU: Tighten integration test journey setup
Across our integration tests we were setting up sessions without the required `journey.optionalPaths` property. This is a required type but was not being set as we were bypassing the `Request` type by saying the `request` could be `any`.

Here we make sure we are using the correct types to refer to `request`, `response` and `next` when setting up the stubbed `validateSessionMiddleware`.

A `getPermittedJourneyForPath` function is created to pull the right values for the `req.session.user.journey` property from the state machine.

A few tests were also corrected in this commit set up the journey on the wrong page for the test. Unless explicitly testing redirects as directed by the state machine, integration test sessions should start in a state where the page being tested is the `nextPath` in the session.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>