AUT-4088: Rename variable passed to template
This is a preliminary step ahead of introducing a new feature flag for the new mfa reset journey.
Currently, the link we display to users is purely based on the supportMfaResetWithIpvReset feature
flag. However, we will soon be updating this to be based on two feature flags, to allow us
to deploy the routes by swicthing on the existing feature flag, while not routing users via the new
journey.

This change anticipates this, by renaming the flag we pass through to the mfa templates. This renaming
makes it clearer that this value is not populated solely by the existing flag.

AUT-4088: rename var auth app