AUT-4088: Use new flag in submission to reset password request
When we send a reset password request to the backend, it currently uses the value of
the allowMfaResetAfterPasswordReset field to determine whether or not to write to the account
modifiers table. If the field is 'true', we do not write an entry to this table, meaning
that a user is free to reset their mfa immediately after a reset password journey.

Previously, this behaviour was solely based on the supportMfaResetViaIpv feature flag.

During the testing phase, we will turn on the supportMfaResetViaIpv feature flag, but keep the
new routeUsersToNewIpvJourney flag off. This will allow us to test connectivity of the new journey
(since we will be able to manually hit the new routes at the relevant part of the journey) without
affecting the journey for normal users using the application.

This commit updates the field in the password reset request to be based on both feature flags being on.

This means that, during the testing phase, both test users and normal users will have account modifier
blocks set on them. This has been done to ensure that there are no changes for normal users during
this phase, since during the testing phase in production we largely want to test connectivity of the
new journey, rather than all the behaviour in detail, and the account modifiers behaviour is not essential
to the new journey.