AUT-4103: Put numbered elements on setup auth app page in correct html element
By putting these items in an ordered list element in html, users accessing the service
via assistive technologies (e.g. screen readers) should be able to better understand
the structure of the page.

This necessitated removing the hard-coded numbers from the steps in the translation
files, as these will be provided by the html elements.

This fixes a failure against this WCAG standard:
https://www.w3.org/WAI/WCAG22/Understanding/info-and-relationships