AUT-3905: Move `CONTENT_IDS` to `contentId.ts`
In advance of some refactoring to the structure of `CONTENT_IDS`, this set of values has been moved to where they are used. The expected refactoring will allow paths to have either a string defined as their default, or a function that will return the ID. This should no longer be considered a set of constants, so moved out of `app.constants.ts`.