AUT-3905: Refactor how contentIds are recorded
Previously we had a set of contentId static values with different variants (e.g. reauth, uplift). It has become apparent that a lot more variants would need to be added in order to consolidate the bulk of contentIds as is, where these variants would only be used in one scenario.

Instead, this is restructured so paths can have simply one string that is the contentId for the path, or the path can have a function with a Request argument that returns the contentId depending on what is in the request. This allows us to create a set of helper functions to use for assessing the request to be used within these functions, allowing dynamic contentIds depending on requests to paths but without creating a confusing situation where `ContentIdVariants` has a large list of nuanced and seldome used variants.