AUT-3905: Consolidate email checking contentIds
Move contentIds from a range of controllers and views used to check email address access to `CONTENT_IDS`. The snapshot update corrects an error where when validation errors are rendered on the same page the contentId is cleared out. The empty contentId added in the controller is to ensure that in an error scenario with a different template that the contentId for the path is not used. This matches the expectation from an existing snapshot test.