AUT-3905: Consolidate creation resend MFA code contentIds
The template holds the hardcoded value used for the contentId of the resend MFA code page for account creation. This is pulled into the consolidated CONTENT_IDS.

Note the controller has a failed attempt to override the contentIds for the page. This is overwritten by the above. Note that the ID `f463a280-31f1-43c0-a2f5-6b46b1e2bb15` is the same ID used for `RESEND_MFA_CODE` when not a reauth journey. This leads me to believe the value here is a mistake and can be removed. Moreover the snapshot tests show this value is not used here.