AUT-4178: Explicitly call session.touch before saving it
Since we started explicitly saving the session, we no longer have a rolling ttl on the redis session. This was an unintended consequence of explicitly calling session.save() to fix a race condition.

Normally, connect redis resets the ttl on each request. However, it does not on session.save() and since after saving the session we make no other changes to the session, the middleware skips updating the ttl at the end of the request. Source: https://github.com/tj/connect-redis\?tab\=readme-ov-file\#ttl

Implementing this will make the session rolling again, therefore extending it each time the user interacts with the application.