AUT-4178: Detect when session not restored from the database
Add a new 'sessionRestored' flag to the session and set this to 'true when the session is first initialised.
Check that this field exists for the session to be vaild in 'sessionIsValid'
If the database ttl expires (despite the aps cookie still being available) then this field will be undefined, indicating that the session has not been restored from the database.  In this case the session expired page is displayed to users as the session has actually expired.

Temporary fix for an issue where the aps cookie expiry time is extended in a rolling manner, whereas the corresponding ttl for the Redis session item is not, which results in errors.