AUT-4178: Add new sessionRestored flag and initialise
Adds a new 'sessionRestored' flag and initialises it when /authenticate is called, which is the front door for all journeys in to Authentication.
A subsequent change will use this flag to determine if the session has actually been restored from the Redis database.

See 9396e656b6d423ce4c4f66faad09a9294acf17e9