AUT-4178: Check required session field 'email' when calling auth-code
Check that the 'email' field is present on the session before handing over to the 'auth-code' route. If the email is not present then route to the standard session error page.

Fixes a second issue where the frontend session in the Redis datastore expires after 1hr, whereas the frontend session cookie (aps) is rolling and can last longer after being extended on each silent login.

The issue manifests in the logs as

TypeError: Cannot read properties of undefined (reading 'toLowerCase') at /app/dist/middleware/account-interventions-middleware.js:13:50

The first issue was resolved by #2705, but that change did not resolve this second issue.

Why the issue happens

The frontend session in the Redis datastore will always expire after 1hr regardless of any silent logins (issue to be resolved later)
The frontend session cookie (aps) has its expiry time extended on every silent login so there is a mismatch between the cookie expiry and the datastore ttl
If a silent login takes place when the datastore item has disappeared then the frontend session will not be correctly restored to what it was before and some fields will (obviously) be missing
This issue is caused by the 'email' field not being set in the session. You can see why here, if email is not in the session then it is set to 'undefined.
Email is only set on the session by enter-email-controller (1) (2), so if email is no longer on the session, or if a journey does not go through that page (such as silent login), then the email will not be populated and an error will happen on auth-code

The fix

This is a temporary fix to display the standard session expired error page rather than a generic error page in this scenario as the session has actually expired. Further work may be done to make the session rolling.