AUT-4173: Create redirect page to update journey
The users journey needs updating in the session to reflect the pages the user is allowed to be on when clicking the link to `/enter-email-create` and going back to `/sign-in-or-create`.

Following the pattern used for password reset, a new `GET /enter-email-create-request` route is added that updates the session to allow the user to both be on and perform the actions on the `/enter-email-create` page.

The `allowAndPersistUserJourneyMiddleware` is also used on the `/sign-in-or-create` page to make sure if a user clicks the back button they are back in the right state to go down the sign in journey again, and vice versa.