AUT-4150: Replace `asyncHandler` with `express-async-errors`
Following an incident that caused FE downtime due to uncaught async error we are changing our approach to handling them.

Previously we would wrap async controllers with `asyncHandler`, but this was prone to being a forgotten step.

Here we utilise a new package to patch express itself to handle these async errors.

Express v5 promises to handle these themselves, but we've not got to upgrading yet.