BAU: Correct authdev signin URLs in env script
Since the authdev environments have moved AWS accounts and the old `*.sandpit.*` domains have been deactivated we've been unable to run the frontend locally targeting the authdev environments.

This updates the env build script to statically create the audience values needed for all environments.

This is needed because the state file needed for the authdev envs reference the incorrect values, which are still required at the moment.