AUT-1416: Create `upsertDefaultSmsMfaMethod`
Shortly use of `session.user.phoneNumber` and `session.user.redactedPhoneNumber` will be replaced with `session.user.mfaMethods` which can contain multiple MFA methods. We are explicitly not handling `BACKUP` `SMS` or any `AUTH_APP` within `mfaMethods` at the moment. Modifications will be made when required, not before, to keep this migration pretty straightforward.

This new function is to be used to replace where `phoneNumber` and `redactedPhoneNumber` are currently set in the user session.