AUT-1416: Create temporary `migrateMfaSessionStorageMiddleware`
Shortly use of `session.user.phoneNumber` and `session.user.redactedPhoneNumber` will be replaced with `session.user.mfaMethods` which can contain multiple MFA methods. We are explicitly not handling `BACKUP` `SMS` or any `AUTH_APP` within `mfaMethods` at the moment. Modifications will be made when required, not before, to keep this migration pretty straightforward.

Here we create a middleware so every time a page is requested and the user has a session we ensure that any `phoneNumber` or `redactedPhoneNumber` is migrated over to the new `mfaMethods` structure.

If `mfaMethods` already exists in the user session we do not overwrite this.

This middleware is only to facilitate the short period of time where a user starts and finishes their journey while the deployment of this change happens at the same time, meaning new code that looks for data in `mfaMethods` will find data even though it wasn't originally set in `mfaMethods`.