AUT-1416: Update `MfaMethod` type
Now that the `frontend-api` implementation of the `mfaMethods` property has been completed, here the types of the return values are updated to reflect that the `id` will be returned and `priority` is used rather than `priorityIdentifier`.

Note `id` is marked as optional as it is currently used for holding partial MFAs while the full method management work is implemented. This may change to a strict required later.