AUT-4272: Switch to use a single csrf token per session
Refreshing per request was too fragile - we had issues when an optional script was not found causing the session token to refresh meaning the token in the form submission did not match (causing a 403). This means that with per request refresh if we make a mistake with an optional script and it 404s then the entire app is broken. This may also be problematic with slow page loads and double submits.

Note that having a token per session mimics the previous implementation with `csurf` and is simpler than per request. This creates a session even when not authenticated, so works with the anonymous pages (e.g., cookies, contact-us).

Other reference:
https://security.stackexchange.com/a/22936